refactor(login): replaced loginSection with userEntry
