--- components/loginSection/index.tsx
@@ -1,12 +1,0 @@
-import styles from './loginSection.module.scss'
-import Login from './login'
-
-const LoginSection = () => {
-  return (
-    <div className={styles.main_container}>
-      <Login />
-    </div>
-  )
-}
-
-export default LoginSection
