--- components/loginSection/loginSection.module.scss
@@ -1,11 +1,0 @@
-$border: 1px solid black;
-
-.login_container {
-  border: $border;
-  width: fit-content;
-  padding: 1rem;
-}
-
-.email_input, .password_input {
-  display: block;
-}
